fix minor issues in document CRUD, added basic tags

+++ src/pages/documents.tsx
@@ -1,13 +1,38 @@
-import React, { useEffect, useState } from 'react';
+import React, { useEffect, useState, useCallback } from 'react';
+import dynamic from "next/dynamic";
 import { useSupabaseClient } from "@supabase/auth-helpers-react";
 import { useRouter } from 'next/router';
-import ReactQuill from 'react-quill';
 import { convert } from 'html-to-text';
 import { Configuration, OpenAIApi } from 'openai'
 
 import BackButton from '@/components/BackButton';
+import Loader from '@/components/Loader';
 
 import 'react-quill/dist/quill.snow.css';
+import 'quill-mention/dist/quill.mention.css';
+
+const QuillMention = dynamic(import('quill-mention'), { ssr: false });
+const ReactQuill = dynamic(import('react-quill'), { ssr: false });
+
+// ReactQuill.Quill.register('modules/mentions', QuillMention)
+
+type Documents = {
+	id: number
+	content: string
+	metadata?: { id: number }
+	embedding: [number]
+	html_string: string
+}
+
+const atValues = [
+	{ id: 1, value: "At Value 1" },
+	{ id: 2, value: "At Value 2" },
+];
+const hashValues = [
+	{ id: 3, value: "Hash Value 1" },
+	{ id: 4, value: "Hash Value 2" },
+	{ id: 5, value: "Story" },
+];
 
 const Documents = () => {
 	const router = useRouter();
@@ -15,8 +40,9 @@ const Documents = () => {
 	const configuration = new Configuration({ apiKey: process.env.NEXT_PUBLIC_API_KEY })
 	const openAi = new OpenAIApi(configuration)
 
-	const [documents, setDocuments] = useState([]);
-	const [selectedDocument, setSelectedDocument] = useState();
+	const [loading, setLoading] = useState(false);
+	const [documents, setDocuments] = useState<Array<Documents>>([]);
+	const [selectedDocument, setSelectedDocument] = useState<Documents>();
 	const [value, setValue] = useState("");
 	const [confirmDelete, setConfirmDelete] = useState(false);
 
@@ -29,14 +55,42 @@ const Documents = () => {
 			})
 	}, [])
 
-	const goToDocument = (id) => {
+	useEffect(() => {
+		if (router.query.id && documents.length > 0) {
+			const foundDoc = documents.find((d) => d.id === parseInt(router.query.id));
+			if (foundDoc) {
+				setSelectedDocument(foundDoc);
+				setValue(foundDoc.html_string);
+			}
+		}
+	}, [router.query, documents])
+
+	useEffect(() => {
+		window.addEventListener('mention-hovered', (event) => { console.log('hovered: ', event) }, false);
+		window.addEventListener('mention-clicked', (event) => { console.log('clicked: ', event) }, false);
+		return () => {
+			window.removeEventListener('mention-hovered', (event) => { console.log('hovered: ', event) });
+			window.removeEventListener('mention-clicked', (event) => { console.log('clicked: ', event) });
+		}
+	}, [])
+
+
+	const goToDocument = (id: number | null) => {
+		if (!id) {
+			setSelectedDocument(undefined);
+			setValue("");
+			return;
+		}
 		const document = documents.find(d => id === d.id);
-		setSelectedDocument(document);
-		setValue(document.html_string);
+		if (document) {
+			setSelectedDocument(document);
+			setValue(document.html_string);
+		}
 	}
 
 	const onDeleteDocument = async () => {
-		if (confirmDelete) {
+		if (confirmDelete && selectedDocument) {
+			setLoading(true);
 			const { data, error } = await supabase
 				.from('documents')
 				.delete()
@@ -44,21 +98,28 @@ const Documents = () => {
 				.select()
 
 			setConfirmDelete(false);
-			setSelectedDocument(false);
+			setSelectedDocument(undefined);
 			setValue("");
 
-			const newDocuments = [...documents];
-			newDocuments.splice(newDocuments.findIndex(d => d.id === data[0].id), 1)
-			setDocuments(newDocuments);
+			if (data) {
+				const newDocuments = [...documents];
+				newDocuments.splice(newDocuments.findIndex(d => d.id === data[0].id), 1)
+				setDocuments(newDocuments);
+			}
 
-			if (error) return console.error("Something went wrong!", error);
+			if (error) {
+				console.error("Something went wrong!", error);
+				setLoading(false);
+				return;
+			}
+			setLoading(false);
 			return console.log("Success!");
 		}
-
 		return setConfirmDelete(true);
 	}
 
 	const onAddDocument = async () => {
+		setLoading(true);
 		try {
 			const input = value.replace(/\n/g, ' ')
 
@@ -80,13 +141,20 @@ const Documents = () => {
 					.eq('id', selectedDocument.id)
 					.select();
 
-				const newDocuments = [...documents];
-				newDocuments.splice(newDocuments.findIndex(d => d.id === data[0].id), 1, data[0])
-				setDocuments(newDocuments);
+				if (data) {
+					const newDocuments = [...documents];
+					newDocuments.splice(newDocuments.findIndex(d => d.id === data[0].id), 1, data[0])
+					setDocuments(newDocuments);
 
-				setValue("");
-				setSelectedDocument(false);
-				if (error) return console.error("Something went wrong!", error);
+					setValue("");
+					setSelectedDocument(undefined);
+				}
+
+				if (error) {
+					console.error("Something went wrong!", error);
+					setLoading(false);
+					return;
+				}
 			} else {
 				const id = Math.floor(Date.now() / 1000);
 
@@ -103,16 +171,60 @@ const Documents = () => {
 					.insert(insert)
 					.select();
 
-				setDocuments([...documents, data[0]]);
-				setValue("");
+				if (data) {
+					setDocuments([...documents, data[0]]);
+					setValue("");
+				}
 
-				if (error) return console.error("Something went wrong!", error);
+				if (error) {
+					console.error("Something went wrong!", error);
+					setLoading(false);
+					return;
+				}
 			}
+
+			setLoading(false);
 			return console.log("Success!");
 
 		} catch (err) {
 			console.error(err);
 		}
+	}
+
+	const mentionModule = {
+		allowedChars: /^[A-Za-z\sÅÄÖåäö]*$/,
+		mentionDenotationChars: ['@', '#'],
+		source: useCallback(
+			(
+				searchTerm: string,
+				renderItem: (
+					arg0: { id: number; value: string }[] | undefined,
+					arg1: any
+				) => void,
+				mentionChar: string
+			) => {
+				let values;
+
+				if (mentionChar === '@') {
+					values = atValues;
+				} else if (mentionChar === '#') {
+					values = hashValues;
+				}
+
+				if (searchTerm.length === 0) {
+					renderItem(values, searchTerm);
+				} else if (values) {
+					const matches = [];
+					for (let i = 0; i < values.length; i += 1)
+						if (
+							values[i].value.toLowerCase().indexOf(searchTerm.toLowerCase())
+						)
+							matches.push(values[i]);
+					renderItem(matches, searchTerm);
+				}
+			},
+			[]
+		),
 	}
 
 	return (
@@ -128,30 +240,36 @@ const Documents = () => {
 			</div>
 
 			<div className="flex flex-1 items-stretch">
-				<div className='w-100 border-r-2 border-slate-400 p-4'>
-					<h1 className="text-lg font-bold"> Document List</h1>
-					<table className="border-collapse border border-slate-500 w-full">
-						<thead>
-							<tr>
-								<th className="border border-slate-600 bg-stone-500">ID</th>
-								<th className="border border-slate-600 bg-stone-500">Content</th>
-							</tr>
-						</thead>
-						<tbody>
-							{
-								documents.map(d => (
-									<tr onClick={() => goToDocument(d.id)} key={d.id}>
-										<td className="border border-slate-700">{d.id}</td>
-										<td className="border border-slate-700">{d.content.length > 20 ? `${d.content.substring(0, 20)}...` : d.content}</td>
-									</tr>
-								))
-							}
-						</tbody>
-					</table>
+				<div className='w-300 border-r-2 border-slate-400 p-4'>
+					<h1 className="text-lg font-bold mb-3"> Document List</h1>
+					<ul>
+						<li className={`border-l-2 pl-2 py-1 hover:border-cyan-300 hover:text-cyan-300 ${!selectedDocument?.id && "border-cyan-500 text-cyan-500"}`}>
+							<a onClick={() => goToDocument(null)} className="whitespace-nowrap cursor-pointer">
+								{`+ Create New Document`}
+							</a>
+						</li>
+						{
+							documents.map((d) =>
+								<li key={d.id} className={`border-l-2 pl-2 py-1 hover:border-cyan-300 hover:text-cyan-300 ${selectedDocument?.id === d.id && "border-cyan-500 text-cyan-500"}`}>
+									<a onClick={() => goToDocument(d.id)} className="whitespace-nowrap cursor-pointer">
+										<label className='mr-2'>{`${d.id} ~`}</label>
+										{d.content.length > 20 ? `${d.content.substring(0, 20)}...` : d.content}
+									</a>
+								</li>
+							)
+						}
+					</ul>
 				</div>
 				<div className='p-4 w-full'>
-					<h1 className='text-lg font-bold'>Text Editor </h1>
-					<ReactQuill bounds=".quill" theme="snow" value={value} onChange={setValue} />
+					<ReactQuill
+						bounds=".quill"
+						theme="snow"
+						value={value}
+						onChange={setValue}
+						modules={{
+							mention: mentionModule
+						}}
+					/>
 					<div>
 						{
 							selectedDocument &&
@@ -159,9 +277,13 @@ const Documents = () => {
 								confirmDelete ?
 									<div className='float-left block'>
 										<span className='text-red-600'> Are you sure you want to delete? </span>
-										<button className="mt-2 block rounded bg-red-500 px-6 pb-2 pt-2.5 text-xs font-medium uppercase leading-normal text-white shadow-[0_4px_9px_-4px_#3b71ca] transition duration-150 ease-in-out hover:bg-primary-600 hover:shadow-[0_8px_9px_-4px_rgba(59,113,202,0.3),0_4px_18px_0_rgba(59,113,202,0.2)] focus:bg-primary-600 focus:shadow-[0_8px_9px_-4px_rgba(59,113,202,0.3),0_4px_18px_0_rgba(59,113,202,0.2)] focus:outline-none focus:ring-0 active:bg-primary-700 active:shadow-[0_8px_9px_-4px_rgba(59,113,202,0.3),0_4px_18px_0_rgba(59,113,202,0.2)] dark:shadow-[0_4px_9px_-4px_rgba(59,113,202,0.5)] dark:hover:shadow-[0_8px_9px_-4px_rgba(59,113,202,0.2),0_4px_18px_0_rgba(59,113,202,0.1)] dark:focus:shadow-[0_8px_9px_-4px_rgba(59,113,202,0.2),0_4px_18px_0_rgba(59,113,202,0.1)] dark:active:shadow-[0_8px_9px_-4px_rgba(59,113,202,0.2),0_4px_18px_0_rgba(59,113,202,0.1)]"
+										<button className={`mt-2 block rounded bg-red-500 px-6 pb-2 pt-2.5 text-xs font-medium uppercase leading-normal text-white shadow-[0_4px_9px_-4px_#3b71ca] transition duration-150 ease-in-out hover:bg-primary-600 hover:shadow-[0_8px_9px_-4px_rgba(59,113,202,0.3),0_4px_18px_0_rgba(59,113,202,0.2)] focus:bg-primary-600 focus:shadow-[0_8px_9px_-4px_rgba(59,113,202,0.3),0_4px_18px_0_rgba(59,113,202,0.2)] focus:outline-none focus:ring-0 active:bg-primary-700 active:shadow-[0_8px_9px_-4px_rgba(59,113,202,0.3),0_4px_18px_0_rgba(59,113,202,0.2)] dark:shadow-[0_4px_9px_-4px_rgba(59,113,202,0.5)] dark:hover:shadow-[0_8px_9px_-4px_rgba(59,113,202,0.2),0_4px_18px_0_rgba(59,113,202,0.1)] dark:focus:shadow-[0_8px_9px_-4px_rgba(59,113,202,0.2),0_4px_18px_0_rgba(59,113,202,0.1)] dark:active:shadow-[0_8px_9px_-4px_rgba(59,113,202,0.2),0_4px_18px_0_rgba(59,113,202,0.1)] ${loading && "cursor-not-allowed"}`}
 											onClick={onDeleteDocument}>
-											Confirm Delete
+											<span className='flex items-center'>
+												{
+													loading ? <><Loader className="mr-2" /> Processing...</> : "Confirm Delete"
+												}
+											</span>
 										</button>
 									</div>
 									:
@@ -172,16 +294,22 @@ const Documents = () => {
 							)
 						}
 
-						<button className="mt-2 float-right inline-block rounded bg-slate-600 px-6 pb-2 pt-2.5 text-xs font-medium uppercase leading-normal text-white shadow-[0_4px_9px_-4px_#3b71ca] transition duration-150 ease-in-out hover:bg-primary-600 hover:shadow-[0_8px_9px_-4px_rgba(59,113,202,0.3),0_4px_18px_0_rgba(59,113,202,0.2)] focus:bg-primary-600 focus:shadow-[0_8px_9px_-4px_rgba(59,113,202,0.3),0_4px_18px_0_rgba(59,113,202,0.2)] focus:outline-none focus:ring-0 active:bg-primary-700 active:shadow-[0_8px_9px_-4px_rgba(59,113,202,0.3),0_4px_18px_0_rgba(59,113,202,0.2)] dark:shadow-[0_4px_9px_-4px_rgba(59,113,202,0.5)] dark:hover:shadow-[0_8px_9px_-4px_rgba(59,113,202,0.2),0_4px_18px_0_rgba(59,113,202,0.1)] dark:focus:shadow-[0_8px_9px_-4px_rgba(59,113,202,0.2),0_4px_18px_0_rgba(59,113,202,0.1)] dark:active:shadow-[0_8px_9px_-4px_rgba(59,113,202,0.2),0_4px_18px_0_rgba(59,113,202,0.1)]"
+						<button className={`mt-2 float-right inline-block rounded bg-slate-600 px-6 pb-2 pt-2.5 text-xs font-medium uppercase leading-normal text-white shadow-[0_4px_9px_-4px_#3b71ca] transition duration-150 ease-in-out hover:bg-primary-600 hover:shadow-[0_8px_9px_-4px_rgba(59,113,202,0.3),0_4px_18px_0_rgba(59,113,202,0.2)] focus:bg-primary-600 focus:shadow-[0_8px_9px_-4px_rgba(59,113,202,0.3),0_4px_18px_0_rgba(59,113,202,0.2)] focus:outline-none focus:ring-0 active:bg-primary-700 active:shadow-[0_8px_9px_-4px_rgba(59,113,202,0.3),0_4px_18px_0_rgba(59,113,202,0.2)] dark:shadow-[0_4px_9px_-4px_rgba(59,113,202,0.5)] dark:hover:shadow-[0_8px_9px_-4px_rgba(59,113,202,0.2),0_4px_18px_0_rgba(59,113,202,0.1)] dark:focus:shadow-[0_8px_9px_-4px_rgba(59,113,202,0.2),0_4px_18px_0_rgba(59,113,202,0.1)] dark:active:shadow-[0_8px_9px_-4px_rgba(59,113,202,0.2),0_4px_18px_0_rgba(59,113,202,0.1)] ${loading && "cursor-not-allowed"}`}
+							disabled={loading}
 							onClick={onAddDocument}>
-							Submit
+							<span className='flex items-center'>
+								{
+									loading ? <><Loader className="mr-2" /> Processing...</> : "Submit"
+								}
+							</span>
+
 						</button>
 					</div>
 
 				</div>
 			</div>
 
-		</div>
+		</div >
 	)
 }
